refinedment / finished

+++ zone/attack.cpp
@@ -3411,11 +3411,11 @@ void Mob::CommonDamage(Mob* attacker, int &damage, const uint16 spell_id, const 
 			if (shielder[0].shielder_id && spell_id == SPELL_UNKNOWN) {
 				Client* shielder_client = entity_list.GetMob(shielder[0].shielder_id)->CastToClient();
 				if (shielder_client) {
-					float dmg_mod = (75. - shielder[0].shielder_bonus / 100.);
+					float dmg_mod = (75. - (float)shielder[0].shielder_bonus) / 100.;
 					if (dmg_mod < 0.5)
 						dmg_mod = 0.5;
 
-					int shielder_dmg = (int) damage*dmg_mod;
+					int32 shielder_dmg = (int32) (damage*dmg_mod);
 					shielder_client->Damage(attacker, shielder_dmg, spell_id, skill_used, avoidable, buffslot, iBuffTic, special);
 					damage = damage / 2;
 				}
